Выполнить Финальное Задание 9: Добавить библиотеку Json. Добавить Новые классы для создания и обработки HttpServer. Добавить тест для проверки сериализации-десериализации. Добавить тесты для проверки работоспособности работы по Http на каждый метод
