+++ src/http/EpicHandler.java
@@ -1,9 +1,10 @@
-package httpTaskServer;
+package http;
 
 import com.google.gson.Gson;
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
 import manager.TaskManager;
+import task.Epic;
 import task.Subtask;
 
 import java.io.IOException;
@@ -11,11 +12,11 @@ import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
 import java.util.List;
 
-public class SubtaskHandler extends BaseHttpHandler implements HttpHandler {
+public class EpicHandler extends BaseHttpHandler implements HttpHandler {
     private final TaskManager taskManager;
     private final Gson gson;
 
-    SubtaskHandler(TaskManager taskManager, Gson gson) {
+    EpicHandler(TaskManager taskManager, Gson gson) {
         this.taskManager = taskManager;
         this.gson = gson;
     }
@@ -27,23 +28,25 @@ public class SubtaskHandler extends BaseHttpHandler implements HttpHandler {
         String[] pathParts = requestPath.split("/");
         switch (requestMethod) {
             case "GET":
-                if (pathParts.length == 2 && pathParts[1].equals("subtasks")) {
+                if (pathParts.length == 2 && pathParts[1].equals("epics")) {
                     getTasksHandle(exchange, gson);
-                } else if (pathParts.length == 3 && pathParts[1].equals("subtasks")) {
+                } else if (pathParts.length == 3 && pathParts[1].equals("epics")) {
                     getTaskHandle(exchange, gson, pathParts);
+                } else if (pathParts.length == 4 && pathParts[1].equals("epics") && pathParts[3].equals("subtasks")) {
+                    getSubtasksHandle(exchange, gson, pathParts);
                 } else {
                     sendNotFound(exchange, "Method not found");
                 }
                 break;
             case "POST":
-                if (pathParts.length == 2 && pathParts[1].equals("subtasks")) {
+                if (pathParts.length == 2 && pathParts[1].equals("epics")) {
                     postTaskHandle(exchange, gson);
                 } else {
                     sendNotFound(exchange, "Method not found");
                 }
                 break;
             case "DELETE":
-                if (pathParts.length == 3 && pathParts[1].equals("subtasks")) {
+                if (pathParts.length == 3 && pathParts[1].equals("epics")) {
                     deleteTaskHandle(exchange, pathParts);
                 } else {
                     sendNotFound(exchange, "Method not found");
@@ -56,7 +59,7 @@ public class SubtaskHandler extends BaseHttpHandler implements HttpHandler {
 
     private void getTasksHandle(HttpExchange exchange, Gson gson) throws IOException {
         try {
-            List<Subtask> tasks = taskManager.getSubtasks();
+            List<Epic> tasks = taskManager.getEpics();
             String text = gson.toJson(tasks);
             sendText(exchange, text);
         } catch (Exception exp) {
@@ -64,13 +67,25 @@ public class SubtaskHandler extends BaseHttpHandler implements HttpHandler {
         }
     }
 
+    private void getSubtasksHandle(HttpExchange exchange, Gson gson, String[] pathParts) throws IOException {
+        try {
+            int id = Integer.parseInt(pathParts[2]);
+            Epic task = taskManager.getEpic(id);
+            List<Subtask> tasks = taskManager.getEpicSubtask(id);
+            String text = gson.toJson(tasks);
+            sendText(exchange, text);
+        } catch (Exception e) {
+            sendNotFound(exchange, "An error occurred during the request" + e.getMessage());
+        }
+    }
+
     private void getTaskHandle(HttpExchange exchange, Gson gson, String[] pathParts) throws IOException {
         try {
             int id = Integer.parseInt(pathParts[2]);
-            Subtask task = taskManager.getSubtask(id);
+            Epic task = taskManager.getEpic(id);
             sendText(exchange, gson.toJson(task));
         } catch (Exception e) {
-            sendNotFound(exchange, "Subtask Id" + pathParts[2] + "not found");
+            sendNotFound(exchange, "Epic Id" + pathParts[2] + "not found");
         }
     }
 
@@ -78,23 +93,19 @@ public class SubtaskHandler extends BaseHttpHandler implements HttpHandler {
         try {
             InputStream bodyInputStream = exchange.getRequestBody();
             String body = new String(bodyInputStream.readAllBytes(), StandardCharsets.UTF_8);
-            Subtask taskDeserialized = gson.fromJson(body, Subtask.class);
+            Epic taskDeserialized = gson.fromJson(body, Epic.class);
             if (taskDeserialized == null) {
                 sendNotFound(exchange, "Invalid format");
                 return;
             }
             if (taskDeserialized.getId() == 0) {
-                int id = taskManager.addNewSubtask(taskDeserialized);
-                if (id == 0) {
-                    sendNotFound(exchange, "Can't create Subtask");
-                } else {
-                    sendSuccessWithoutBody(exchange);
-                }
+                taskManager.addNewEpic(taskDeserialized);
+                sendSuccessWithoutBody(exchange);
             } else {
-                if (taskManager.getSubtask(taskDeserialized.getId()) == null) {
-                    sendNotFound(exchange, "Subtask Id not found" + taskDeserialized.getId());
+                if (taskManager.getEpic(taskDeserialized.getId()) == null) {
+                    sendNotFound(exchange, "Epic Id not found" + taskDeserialized.getId());
                 } else {
-                    taskManager.updateSubtask(taskDeserialized);
+                    taskManager.updateEpic(taskDeserialized);
                     sendSuccessWithoutBody(exchange);
                 }
             }
@@ -106,11 +117,11 @@ public class SubtaskHandler extends BaseHttpHandler implements HttpHandler {
     private void deleteTaskHandle(HttpExchange exchange, String[] pathParts) throws IOException {
         try {
             int id = Integer.parseInt(pathParts[2]);
-            Subtask task = taskManager.getSubtask(id);
-            taskManager.removeSubtaskForId(task.getId());
-            sendText(exchange, "Task was successfully deleted");
+            Epic task = taskManager.getEpic(id);
+            taskManager.removeEpicForId(task.getId());
+            sendText(exchange, "Epic was successfully deleted");
         } catch (Exception e) {
-            sendNotFound(exchange, "Subtask Id" + pathParts[2] + "not founded");
+            sendNotFound(exchange, "Epic Id" + pathParts[2] + "not founded");
         }
     }
 }
