+++ src/http/HistoryHandler.java
@@ -1,4 +1,4 @@
-package httpTaskServer;
+package http;
 
 import com.google.gson.Gson;
 import com.sun.net.httpserver.HttpExchange;
