+++ src/http/LocalDateTimeAdapter.java
@@ -1,4 +1,4 @@
-package httpTaskServer;
+package http;
 
 import com.google.gson.TypeAdapter;
 import com.google.gson.stream.JsonReader;
