+++ test/httpTaksManager/HttpEpicTest.java
@@ -2,8 +2,8 @@ package httpTaksManager;
 
 import com.google.gson.*;
 import exceptions.NotFoundException;
-import httpTaskServer.HttpTaskServer;
-import httpTaskServer.TaskListTypeToken;
+import http.HttpTaskServer;
+import http.TaskListTypeToken;
 import manager.InMemoryTaskManager;
 import manager.TaskManager;
 import org.junit.jupiter.api.AfterEach;
