+++ test/httpTaksManager/HttpPrioritizedTest.java
@@ -4,8 +4,8 @@ import com.google.gson.Gson;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonParser;
-import httpTaskServer.HttpTaskServer;
-import httpTaskServer.TaskListTypeToken;
+import http.HttpTaskServer;
+import http.TaskListTypeToken;
 import manager.InMemoryTaskManager;
 import manager.TaskManager;
 import org.junit.jupiter.api.AfterEach;
